move Okta wrapper our of server
