+++ pkg/server/server.go
@@ -8,11 +8,14 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/spf13/viper"
+
+	"github.com/cmsgov/easi-app/pkg/okta"
 )
 
 type server struct {
-	router *mux.Router
-	Config *viper.Viper
+	router           *mux.Router
+	Config           *viper.Viper
+	authorizeWrapper func(http.HandlerFunc) http.HandlerFunc
 }
 
 func (s *server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
@@ -22,9 +25,16 @@ func (s *server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 // Serve serves all the handlers
 func Serve(config *viper.Viper) {
 	r := mux.NewRouter()
+	// TODO: We should add some sort of config verifier to make sure these configs exist
+	// They may live in /cmd, but should fail quick on startup
+	authWrapper := okta.NewOktaAuthorizeWrapper(
+		config.GetString("OKTA_CLIENT_ID"),
+		config.GetString("OKTA_ISSUER"),
+	)
 	s := &server{
-		router: r,
-		Config: config,
+		router:           r,
+		Config:           config,
+		authorizeWrapper: authWrapper,
 	}
 	fmt.Print("Serving application on localhost:8080")
 	s.routes()
